matching lots of functions in fighter.c, lots of updates to fighter structs in fighter.h as merging different matches, and updating old matches to the latest fighter context

+++ src/melee/pl/player.c
@@ -721,8 +721,8 @@ void Player_SetFacingDirectionConditional(s32 slot, BOOL b, f32 direction) {
     }
 }
 
-s8 Player_GetCostumeId(s32 slot) {   
-    s8 costume_id;
+u32 Player_GetCostumeId(s32 slot) {   
+    u8 costume_id;
     StaticPlayer* player;
     Player_CheckSlot(slot);
     player = &player_slots[slot];
@@ -737,7 +737,7 @@ void Player_SetCostumeId(s32 slot, s8 costume_id) {
     player->costume_id = costume_id;
 }
 
-s8 Player_GetControllerIndex(s32 slot) {   
+u8 Player_GetControllerIndex(s32 slot) {   
     s8 controller_index;
     StaticPlayer* player;
     Player_CheckSlot(slot);
